+++ src/main/kotlin/global/request/Request.kt
@@ -9,26 +9,33 @@ class Request {
         ContextManager.get< WiseSayingController>("wiseSayingController")!!
     }
 
-    fun work(command: Command){
+    fun work(command: Command):Boolean{
         when (command.command){
             "등록"->{
                 wiseSayingController.createWiseSaying(command)
+                return true
             }
             "목록"->{
                 wiseSayingController.getWiseSayings(command)
+                return true
             }
             "삭제"->{
                 wiseSayingController.deleteWiseSaying(command)
+                return true
             }
             "수정"->{
                 wiseSayingController.modifyWiseSaying(command)
+                return true
             }
             "종료"->{
                 wiseSayingController.finalizeDb()
+                return false
             }
             "빌드"->{
                 wiseSayingController.build()
+                return true
             }
+            else->return true
         }
     }
 }
